Update building script

+++ buildPlayer.py
@@ -10,37 +10,59 @@ elif (os.path.exists(os.path.normpath("C:/flex_sdk_4.6"))):
     flex = ("C:/flex_sdk_4.6")
     exe = ".exe"
 
+MAIN_OUTPUT = "mse.swc"
+JWP_OUTPUT = "bin-release/jwplayer.flash.swf"
+VJS_OUTPUT = "dist/video-js-sr.swf"
+
 debug = "false"
-log_pts = "true"
+verbose = False
+log_pts = "false"
 swfversion = "17"
 targetPlayer = "11.4.0"
 
 def helpParam():
-    print "\nbuildPlayer.py [options]\noptions:\n\t-debug : set debug flag to true\n\t-h print this menu\n"
+    print "\nbuildPlayer.py [options]\noptions:\n\t--debug : set debug flag to true\n\t--log-pts : activate PTS log\n\t-v : verbose mode\n\t-h print this menu\n"
     sys.exit(0)
 
 if (len(sys.argv)>1):
     for i in range(1, len(sys.argv)):
-        if sys.argv[i] == "-debug":
+        if sys.argv[i] == "--debug":
             debug = "true"
             swfversion = "18"
             targetPlayer = "11.5.0"
+        elif sys.argv[i] == "--log-pts":
+            log_pts = "true"
+        elif sys.argv[i] == "-v":
+            verbose = True
         elif sys.argv[i] in ["--help","-h"]:
             helpParam()
         else:
             print "incorrect argument"
             helpParam()
-            
-print "Debug flag = " + debug
-print "LOGGING_PTS = " + log_pts 
-print "-swf-version="+swfversion
-print "-target-player="+targetPlayer
+if verbose:            
+    print "Debug flag = " + debug
+    print "LOGGING_PTS = " + log_pts 
+    print "-swf-version="+swfversion
+    print "-target-player="+targetPlayer
 
 def popenPrint(result):
     result.wait()
-    for line in result.stdout:
-        print(line)
+    if verbose:
+        for line in result.stdout:
+            print(line)
             
+    for line in result.stderr:
+        while line.endswith("\n"):
+            line = line[:-1]
+        if not line == "":
+            line = line.replace("Error:", "\n\033[31mError\033[0m:");
+            line = line.replace("Erreur:", "\n\033[31mErreur\033[0m:");
+            line = line.replace("Warning:", "\n\033[33mWarning\033[0m:");
+            line = line.replace("Avertissement:", "\n\033[33mAvertissement\033[0m:");
+            if line.startswith('\n'):
+                line = line[1:]
+            print(line)
+
 if ('sr-flash' not in os.getcwd()):
     os.chdir(os.path.normpath("sr-flash"))
 
@@ -56,12 +78,12 @@ workerResult = subprocess.Popen([os.path.normpath(flex + "/bin/mxmlc" + exe),
                           "-define+=CONFIG::LOGGING,false",
                           "-define+=CONFIG::LOGGING_PTS,true"],
 
-                          stdout=subprocess.PIPE)
+                          stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
 popenPrint(workerResult)
 #Remove the old library before compiling the new as mxmlc doesn't support compling into non empty folder
-if (os.path.exists("mse.swc")):
-    os.remove("mse.swc")
+if (os.path.exists(MAIN_OUTPUT)):
+    os.remove(MAIN_OUTPUT)
 #compile the library
 libResult = subprocess.Popen([os.path.normpath(flex + "/bin/compc" + exe),
                           os.path.normpath("-compiler.source-path=."),
@@ -74,16 +96,18 @@ libResult = subprocess.Popen([os.path.normpath(flex + "/bin/compc" + exe),
                           "-directory=false",
                           "-include-sources",
                           os.path.normpath("com/streamroot/StreamrootMSE.as"),
-                          "-output=mse.swc"], stdout=subprocess.PIPE)
+                          "-output=" + MAIN_OUTPUT], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 popenPrint(libResult)
-
+if not os.path.exists(MAIN_OUTPUT):
+    print "\n\033[31mBuild failed\033[0m"
+    sys.exit(0)
 #Moving the new library to the jwplayer and videojs folder
 if (os.path.exists(os.path.normpath("../streamroot-videojs/src/com/streamroot/mse.swc"))):
     os.remove("../streamroot-videojs/src/com/streamroot/mse.swc")
-shutil.copyfile(os.path.normpath("mse.swc"),os.path.normpath("../streamroot-videojs/src/com/streamroot/mse.swc"))
+shutil.copyfile(os.path.normpath(MAIN_OUTPUT),os.path.normpath("../streamroot-videojs/src/com/streamroot/mse.swc"))
 if (os.path.exists(os.path.normpath("../streamroot-jwplayer/src/flash/com/streamroot/mse.swc"))):
     os.remove(os.path.normpath("../streamroot-jwplayer/src/flash/com/streamroot/mse.swc"))
-shutil.copyfile(os.path.normpath("mse.swc"),os.path.normpath("../streamroot-jwplayer/src/flash/com/streamroot/mse.swc"))
+shutil.copyfile(os.path.normpath(MAIN_OUTPUT),os.path.normpath("../streamroot-jwplayer/src/flash/com/streamroot/mse.swc"))
 
 #Moving to the jwplayer folder and compiling jwplayer
 os.chdir(os.path.normpath("../streamroot-jwplayer"))
@@ -99,15 +123,18 @@ jwpResult = subprocess.Popen([os.path.normpath(flex + "/bin/mxmlc" + exe),
                           "-swf-version="+swfversion+"",
                           "-debug="+debug+"",
                           "-use-network=false",
-                          os.path.normpath("-output=bin-release/jwplayer.flash.swf"),
+                          os.path.normpath("-output="+ JWP_OUTPUT),
                           "-compiler.optimize=true",
                           "-compiler.omit-trace-statements=true",
                           "-warnings=false",
                           "-define+=JWPLAYER::version,'6.11.20141118195350280'",
-                          "-define+=CONFIG::debugging,"+debug+""], stdout=subprocess.PIPE)
+                          "-define+=CONFIG::debugging,"+debug+""], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 popenPrint(jwpResult)
-shutil.copy2(os.path.normpath("bin-release/jwplayer.flash.swf"),os.path.normpath("../sr-client-last/player_wrapper/jwplayer-wrapper/dist/6.8/jwplayer.srflash.swf"))
-
+shutil.copy2(os.path.normpath(JWP_OUTPUT),os.path.normpath("../sr-client-last/player_wrapper/jwplayer-wrapper/dist/6.8/jwplayer.srflash.swf"))
+if not os.path.exists(JWP_OUTPUT):
+    print "\n\033[31mBuild failed\033[0m"
+    sys.exit(0)
+    
 #Moving to the videojs folder and compiling videojs
 os.chdir(os.path.normpath("../streamroot-videojs"))
 vjsResult = subprocess.Popen([os.path.normpath(flex +"/bin/mxmlc" + exe),
@@ -122,11 +149,15 @@ vjsResult = subprocess.Popen([os.path.normpath(flex +"/bin/mxmlc" + exe),
                           "-debug="+debug+"",
                           "-use-network=false",
                           "-static-link-runtime-shared-libraries=true",
-                          os.path.normpath("-output=dist/video-js-sr.swf"),
+                          os.path.normpath("-output=" + VJS_OUTPUT),
                           "-compiler.optimize=true",
                           "-compiler.omit-trace-statements=true",
                           "-warnings=false",
-                          "-define+=CONFIG::version,\"'4.2.2'\""], stdout=subprocess.PIPE)
+                          "-define+=CONFIG::version,\"'4.2.2'\""], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 popenPrint(vjsResult)
-shutil.copy2(os.path.normpath("dist/video-js-sr.swf"),os.path.normpath("../sr-client-last/player_wrapper/video-js-wrapper/dist/video-js-sr.swf"))
-print "Successfully built the flash engine swfs! Nice Job!"
+shutil.copy2(os.path.normpath(VJS_OUTPUT),os.path.normpath("../sr-client-last/player_wrapper/video-js-wrapper/dist/video-js-sr.swf"))
+if not os.path.exists(VJS_OUTPUT):
+    print "\n\033[31mBuild failed\033[0m"
+    sys.exit(0)
+
+print "\n\033[32mBuild successful\033[0m"
